services/auth/requirements.txt
coverage==7.8.2
fastapi==0.115.12
gunicorn==23.0.0
httpx==0.28.1
pydantic-settings==2.9.1
pytest==8.3.5
pytest-asyncio==1.0.0
pytest-cov==6.1.1
python-jose[cryptography]==3.5.0
redis==6.2.0
uvicorn==0.34.2